+++ packages/sdk-metrics/src/state/WritableMetricStorage.ts
@@ -13,11 +13,14 @@ import type { AttributeHashMap } from './HashMap';
  * An interface representing SyncMetricStorage with type parameters removed.
  */
 export interface WritableMetricStorage {
+  /** Whether this storage has an attribute processor that needs context. */
+  hasAttributeProcessor: boolean;
+
   /** Records a measurement. */
   record(
     value: number,
     attributes: Attributes,
-    context: Context,
+    context: Context | undefined,
     recordTime: number
   ): void;
 }